Change parameter type from int to Long for product methods

+++ src/main/java/com/ProyectoDeAula5/Proyecto5/service/ProductoService.java
@@ -34,7 +34,7 @@ public class ProductoService {
         productoRepository.deleteById(id);
     }
 
-    public void actualizarInventario(int codigo, int cantidad) {
+    public void actualizarInventario(Long codigo, int cantidad) {
         log.info("Actualizando productos");
         Producto producto = productoRepository.findByCodigo(codigo);
         if (producto != null) {
@@ -47,7 +47,7 @@ public class ProductoService {
         }
     }
 
-    public Producto obtenerPorCodigo(int codigo) {
+    public Producto obtenerPorCodigo(Long codigo) {
         return productoRepository.findByCodigo(codigo);
     }
